refactor: 957 - list component
