+++ src/shared/ui/list/index.ts
@@ -1,1 +1,2 @@
-export { List } from './list';
+export { List } from './ui/list/list';
+export { ListItem } from './ui/list-item/list-item';
