+++ src/shared/ui/list/ui/list-item/list-item.module.scss
