+++ src/shared/ui/list/ui/list/list.module.scss
@@ -10,17 +10,6 @@
   letter-spacing: 0.5px;
   list-style-type: none;
 
-  &.compact {
-    gap: 0;
-    letter-spacing: 0;
-
-    @include media-laptop {
-      padding-bottom: 14px;
-      font-size: $font-size-s;
-      line-height: 24px;
-    }
-  }
-
   &.marked {
     list-style-type: disc;
   }
